requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.87.0
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
typing_extensions==4.4.0
uvicorn==0.19.0
